requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-admin-rangefilter==0.6.3
django-cors-headers==3.5.0
django-multiselectfield==0.1.12
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
djoser==2.0.5
drf-yasg==1.20.0
freeze==3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
qrcode==6.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
